Edit QuizException, QuizService, QuizDAO added QuizServiceImplTest

+++ src/main/java/ua/netcracker/netcrackerquizb/service/Impl/QuizServiceImpl.java
@@ -1,14 +1,12 @@
 package ua.netcracker.netcrackerquizb.service.Impl;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.netcracker.netcrackerquizb.dao.QuizDAO;
-import ua.netcracker.netcrackerquizb.exception.DAOLogicException;
-import ua.netcracker.netcrackerquizb.exception.QuizDoesNotExistException;
-import ua.netcracker.netcrackerquizb.exception.UserDoesNotExistException;
+import ua.netcracker.netcrackerquizb.exception.*;
 import ua.netcracker.netcrackerquizb.model.Quiz;
 import ua.netcracker.netcrackerquizb.model.QuizType;
-import ua.netcracker.netcrackerquizb.model.User;
 import ua.netcracker.netcrackerquizb.model.impl.QuizImpl;
 import ua.netcracker.netcrackerquizb.service.QuizService;
 
@@ -16,25 +14,62 @@ import java.math.BigInteger;
 import java.sql.Date;
 import java.util.List;
 
+import static ua.netcracker.netcrackerquizb.exception.MessagesForException.*;
+
+
 @Service
 public class QuizServiceImpl implements QuizService {
+
+    private static final Logger log = Logger.getLogger(QuizServiceImpl.class);
 
     @Autowired
     private QuizDAO quizDAO;
 
     @Override
-    public Quiz createQuiz(Quiz quiz) throws DAOLogicException, UserDoesNotExistException {
-        return quizDAO.createQuiz(quiz);
+    public Quiz buildNewQuiz(String title, String description, QuizType quizType, BigInteger userId) throws QuizException, DAOLogicException {
+        title = title.trim();
+        description = description.trim();
+        try {
+            if (quizDAO.existQuizByDescription(description)) {
+                log.info(QUIZ_ALREADY_EXISTS);
+                throw new QuizException(QUIZ_ALREADY_EXISTS);
+            }
+            Quiz quiz = QuizImpl.QuizBuilder()
+                    .setTitle(title)
+                    .setDescription(description)
+                    .setQuizType(quizType)
+                    .setCreationDate(new Date(System.currentTimeMillis()))
+                    .setCreatorId(userId)
+                    .build();
+
+            return quizDAO.createQuiz(quiz);
+
+        } catch (DAOLogicException | UserDoesNotExistException e) {
+            log.info(DAO_LOGIC_EXCEPTION + " in buildNewQuiz()");
+            throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
+        }
+
     }
 
     @Override
     public void updateQuiz(Quiz updatedQuiz) throws QuizDoesNotExistException, DAOLogicException {
-        quizDAO.updateQuiz(updatedQuiz);
+        Quiz quizFromDAO = quizDAO.getQuizById(updatedQuiz.getId());
+        if (quizFromDAO != null) {
+            quizDAO.updateQuiz(updatedQuiz);
+        } else {
+            throw new QuizDoesNotExistException(QUIZ_NOT_FOUND_EXCEPTION);
+        }
     }
 
     @Override
     public void deleteQuiz(Quiz quiz) throws QuizDoesNotExistException, DAOLogicException {
-        quizDAO.deleteQuiz(quiz);
+        Quiz quizFromDAO = quizDAO.getQuizById(quiz.getId());
+        if (quizFromDAO != null) {
+            quizDAO.deleteQuiz(quiz);
+        } else {
+            throw new QuizDoesNotExistException(QUIZ_NOT_FOUND_EXCEPTION);
+        }
+
     }
 
     @Override
@@ -53,18 +88,11 @@ public class QuizServiceImpl implements QuizService {
     }
 
     @Override
-    public List<Quiz> getQuizzesByTitle(String title) throws QuizDoesNotExistException, DAOLogicException {
+    public List<Quiz> getQuizzesByTitle(String title) throws QuizDoesNotExistException, DAOLogicException, QuizException {
+        if (title.isBlank()) {
+            throw new QuizException(EMPTY_TITLE);
+        }
         return quizDAO.getQuizzesByTitle(title);
     }
 
-    @Override
-    public Quiz buildNewQuiz(String title, String description, QuizType quizType, User user) {
-        return QuizImpl.QuizBuilder()
-                .setTitle(title)
-                .setDescription(description)
-                .setQuizType(quizType)
-                .setCreationDate(new Date(System.currentTimeMillis()))
-                .setCreatorId(user.getId())
-                .build();
-    }
 }
